created header file, and moved files to relevant position.
